% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhs_calibration_HS_2.R,
%   R/sndet_calibration_HS_2.R
\name{fun_calibration_HS_2}
\alias{fun_calibration_HS_2}
\alias{fun_calibration_HS_2}
\title{Calibrate Snow Height signal using virtual surveys (HS = 0 cm at end of winter season)}
\usage{
fun_calibration_HS_2(DATA, FILE_NAME, PATH_SURVEYS)

fun_calibration_HS_2(DATA, FILE_NAME, PATH_SURVEYS)
}
\arguments{
\item{DATA}{zoo data}

\item{FILE_NAME}{file in input folder to process (hourly aggregated)}

\item{PATH_SURVEYS}{Path of virtual snow surveys}

\item{DATA}{zoo data}

\item{FILE_NAME}{file in input folder to process (hourly aggregated)}

\item{PATH_SURVEYS}{Path of virtual snow surveys}
}
\description{
Calibrate Snow Height signal using virtual surveys (HS = 0 cm at end of winter season)

Calibrate Snow Height signal using virtual surveys (HS = 0 cm at end of winter season)
}
