% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhs_moving_average.R
\name{fun_moving_average}
\alias{fun_moving_average}
\title{Function that apply a moving average filter}
\usage{
fun_moving_average(DATA, PERIOD_LENGTH)
}
\arguments{
\item{DATA}{zoo data}

\item{PERIOD_LENGTH}{moving average window}
}
\description{
Function that apply a moving average filter
}
