% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sndet_phar_snow_detection.R
\name{fun_phar_snow_detection}
\alias{fun_phar_snow_detection}
\title{Snow presence derived from Phar sensor. Little radiation that hit soil and little fraction from incoming radiation suggest the presence of snow near the station}
\usage{
fun_phar_snow_detection(PAR_UP, PAR_DOWN, RATIO_THRESHOLD = 0.1,
  PAR_SOIL_THRESHOLD = 75)
}
\arguments{
\item{PAR_UP}{column of zoo data correspondig to PAR_Up (Photosintetically Active Radiation, at 2 meters)}

\item{PAR_DOWN}{column of zoo data correspondig to PAR_Down (Photosintetically Active Radiation, at soil level )}

\item{RATIO_THRESHOLD}{Threshold of fraction of incoming PAR. Suggested value 0.1}

\item{PAR_SOIL_THRESHOLD}{Threshold of PAR at soil level}
}
\description{
Snow presence derived from Phar sensor. Little radiation that hit soil and little fraction from incoming radiation suggest the presence of snow near the station
}
