% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sndet_dygraphs_snow_detection.R
\name{fun_plot_models}
\alias{fun_plot_models}
\title{Function that plot snow presence using three models: only PAR, only Soil Temperature, combination of PAR and Soil Temperature without Snow Height}
\usage{
fun_plot_models(FILE, ST_MODEL, PAR_MODEL, MODEL)
}
\arguments{
\item{FILE}{name of file to plot for title}

\item{ST_MODEL}{snow presence using only some checks on Soil Temperature}

\item{PAR_MODEL}{snow presence using only some checks on Photosintetically Active Radiation}

\item{MODEL}{snow presence using the combination of Soil Temperature and PAR}
}
\description{
Function that plot snow presence using three models: only PAR, only Soil Temperature, combination of PAR and Soil Temperature without Snow Height
}
