% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhs_rate.R
\name{fun_rate}
\alias{fun_rate}
\title{Function that exclude data having a rate too high (increase/decrease too fast)}
\usage{
fun_rate(DATA, VARIABLE, RATE)
}
\arguments{
\item{DATA}{zoo data}

\item{VARIABLE}{variable to apply filter of increasing/decresaing rate}
}
\description{
Function that exclude data having a rate too high (increase/decrease too fast)
}
