% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esqc_Read_data_metadata.R
\name{fun_read_output_data}
\alias{fun_read_output_data}
\title{Function that import and prepare data in the proper format for ESOLIP_QC}
\usage{
fun_read_output_data(PATH, FILE)
}
\arguments{
\item{PATH}{path of input folder}

\item{FILE}{file in input folder to process (hourly aggregated)}
}
\description{
Function that import and prepare data in the proper format for ESOLIP_QC
}
