% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhs_savitzky_golay_filter.R
\name{fun_savitzky_golay}
\alias{fun_savitzky_golay}
\title{Function that apply a savitzky_golay. See https://cran.r-project.org/web/packages/signal/signal.pdf}
\usage{
fun_savitzky_golay(DATA, FILTER_ORDER, FILTER_LENGTH)
}
\arguments{
\item{DATA}{zoo data}

\item{FILTER_ORDER}{filter order}

\item{FILTER_LENGTH}{filter length (must be odd)}
}
\description{
Function that apply a savitzky_golay. See https://cran.r-project.org/web/packages/signal/signal.pdf
}
