% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sndet_snow_detection.R
\name{fun_snow_detection}
\alias{fun_snow_detection}
\title{Function that merge snow presence from Soil Temp and snow presence from PAR to improve results}
\usage{
fun_snow_detection(SOIL_TEMP_SNOW, PHAR_SNOW)
}
\arguments{
\item{DATA}{zoo data}

\item{FILE}{variable to clean excluding outliers}
}
\description{
Function that merge snow presence from Soil Temp and snow presence from PAR to improve results
}
