% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sndet_soil_temp_snow_detection.R
\name{fun_soil_temp_snow_detection}
\alias{fun_soil_temp_snow_detection}
\title{Snow presence derived from Soil Temperature sensor. Soil Temperature during snow presence has generally a little daily mean and small daily amplitude.
# MEAN_ST_THRESHOLD = 3.5               # Suggested value: 3.5 . Units: deg C (daily mean)}
\usage{
fun_soil_temp_snow_detection(SOIL_TEMPERATURE, MEAN_ST_THRESHOLD = 3.5,
  AMPLITUDE_ST_THRESHOLD = 3)
}
\arguments{
\item{SOIL_TEMPERATURE}{column of zoo data correspondig to PAR_Up (Photosintetically Active Radiation, at 2 meters)}

\item{MEAN_ST_THRESHOLD}{Threshold of daily mean of soil temperature. Suggested value 3.5}

\item{MEAN_ST_THRESHOLD}{Threshold of daily amplitude of soil temperature. Suggested value 3}
}
\description{
Snow presence derived from Soil Temperature sensor. Soil Temperature during snow presence has generally a little daily mean and small daily amplitude.
# MEAN_ST_THRESHOLD = 3.5               # Suggested value: 3.5 . Units: deg C (daily mean)
}
