\name{SnowSeason_analysis-package}
\alias{SnowSeason_analysis-package}
\alias{SnowSeason_analysis}
\docType{package}
\title{
\packageTitle{SnowSeason_analysis}
}
\description{
\packageDescription{SnowSeason_analysis}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SnowSeason_analysis}
\packageIndices{SnowSeason_analysis}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SnowSeason_analysis}

Maintainer: \packageMaintainer{SnowSeason_analysis}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
